Fix all mobs being pigs in 1.17/1.18
