+++ NMS_Master/src/main/java/simplepets/brainsynder/nms/entity/EntityBase.java
@@ -44,20 +44,6 @@ public class EntityBase extends Mob {
         return user;
     }
 
-    @Override
-    public Packet<?> getAddEntityPacket() {
-        try {
-            ClientboundAddMobPacket packet = new ClientboundAddMobPacket(this);
-            Field type = packet.getClass().getDeclaredField(VersionTranslator.getEntityTypeVariable());
-            type.setAccessible(true);
-            type.set(packet, Registry.ENTITY_TYPE.getId(originalEntityType));
-            return packet;
-        } catch (Exception ex) {
-            ex.printStackTrace();
-        }
-        return new ClientboundAddEntityPacket(this, originalEntityType, 0, new BlockPos(getX(), getY(), getZ()));
-    }
-
     EntityType<? extends Mob> getEntityType(EntityType<? extends Mob> originalType)  {
         try {
             Field field = EntityType.class.getDeclaredField(VersionTranslator.ENTITY_FACTORY_FIELD);
