+++ NMS_Master/src/main/java/simplepets/brainsynder/nms/entity/EntityPet.java
@@ -3,7 +3,12 @@ package simplepets.brainsynder.nms.entity;
 import lib.brainsynder.nbt.StorageTagCompound;
 import lib.brainsynder.sounds.SoundMaker;
 import lib.brainsynder.utils.Colorize;
+import net.minecraft.core.BlockPos;
+import net.minecraft.core.Registry;
 import net.minecraft.nbt.CompoundTag;
+import net.minecraft.network.protocol.Packet;
+import net.minecraft.network.protocol.game.ClientboundAddEntityPacket;
+import net.minecraft.network.protocol.game.ClientboundAddMobPacket;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.damagesource.DamageSource;
 import net.minecraft.world.entity.EntityType;
@@ -196,6 +201,20 @@ public abstract class EntityPet extends EntityBase implements IEntityPet {
                 + Colorize.translateBungeeHex(event.getSuffix());
         getBukkitEntity().setCustomNameVisible(ConfigOption.INSTANCE.PET_TOGGLES_SHOW_NAMES.getValue());
         getBukkitEntity().setCustomName(petName);
+    }
+
+    @Override
+    public Packet<?> getAddEntityPacket() {
+        try {
+            ClientboundAddMobPacket packet = new ClientboundAddMobPacket(this);
+            Field type = packet.getClass().getDeclaredField(VersionTranslator.getEntityTypeVariable());
+            type.setAccessible(true);
+            type.set(packet, VersionTranslator.useInteger() ? Registry.ENTITY_TYPE.getId(originalEntityType) : originalEntityType);
+            return packet;
+        } catch (Exception ex) {
+            ex.printStackTrace();
+        }
+        return new ClientboundAddEntityPacket(this, originalEntityType, 0, new BlockPos(getX(), getY(), getZ()));
     }
 
     @Override
